+++ src/app/src/index.tsx
@@ -7,16 +7,21 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { ChakraProvider } from '@chakra-ui/react';
+import { Provider } from 'react-redux';
 
 import App from './App';
 import theme from './theme';
+import { store } from './store';
+
 
 const root = createRoot(document.getElementById('root')!);
 
 root.render(
     <React.StrictMode>
+        <Provider store={store}>
         <ChakraProvider theme={theme}>
             <App />
         </ChakraProvider>
+        </Provider>
     </React.StrictMode>
 );